Refactor terminal ops so we only manipulate `/dev/tty`
